+++ src/controls/text.ts
@@ -302,7 +302,8 @@ export default class WUPTextControl<
           mask: none;
         }
         :host:focus-within button[clear] {
-          display: block;
+          display: inline-block;
+          opacity: 1;
         }
         :host button[clear=back] {
           --ctrl-icon-img: var(--wup-icon-back);
@@ -324,7 +325,8 @@ export default class WUPTextControl<
         }
         @media (hover: hover) and (pointer: fine) {
           :host:hover button[clear] {
-            display: block;
+            display: inline-block;
+            opacity: 1;
           }
           :host button[clear]:hover:before {
             content: "";
@@ -770,5 +772,3 @@ export default class WUPTextControl<
 customElements.define(tagName, WUPTextControl);
 // todo example how to create bult-in dropdown before the main input (like phone-number with ability to select countryCode)
 // gotInput > setMask > parseValue >... setValue ....> toString > setInput > setMask
-
-// todo display of btnClear changes control size